StockData bought total value seems to be rounding
Was using getInt for StockPrice when it should be getDouble

+++ app/src/main/java/br/com/carteira/fragment/BaseFragment.java
@@ -474,7 +474,7 @@ public abstract class BaseFragment extends Fragment {
                 switch (currentType){
                     case Constants.Type.BUY:
                         quantityTotal += STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_QUANTITY));
-                        valueTotal += STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_QUANTITY))*STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_PRICE));
+                        valueTotal += STQueryCursor.getInt(STQueryCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_QUANTITY))*STQueryCursor.getDouble(STQueryCursor.getColumnIndex(PortfolioContract.StockTransaction.COLUMN_PRICE));
                         mediumPrice = valueTotal/quantityTotal;
                         break;
                     case Constants.Type.SELL:
